quickstart/part-4/pyproject.toml
[project]
name = "summarize"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Cully Wakelin", email = "cully.wakelin@resonatehq.io" }
]
dependencies = [
    "resonate-sdk>=0.3.4",
    "flask>=3.0.3",
    "charset-normalizer>=3.4.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[project.scripts]
    "gateway" = "summarize.gateway:main"
    "app" = "summarize.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "mypy>=1.13.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/summarize"]

[tool.rye.scripts]
check-types = "dmypy run --timeout 60 -- src"
